Implement common interface for Secure and Basic client request response channel.
