+++ src/Vlingo.Wire.Tests/Fdx/Bidirectional/SocketRequestResponseChannelTest.cs
@@ -26,7 +26,7 @@ namespace Vlingo.Wire.Tests.Fdx.Bidirectional
         private static int _testPort = 37471;
 
         private readonly MemoryStream _buffer;
-        private readonly ClientRequestResponseChannel _client;
+        private readonly IClientRequestResponseChannel _client;
         private readonly TestResponseChannelConsumer _clientConsumer;
         private readonly IServerRequestResponseChannel _server;
         private readonly TestRequestChannelConsumer _serverConsumer;
@@ -236,7 +236,7 @@ namespace Vlingo.Wire.Tests.Fdx.Bidirectional
             
             _clientConsumer = new TestResponseChannelConsumer();
             
-            _client = new ClientRequestResponseChannel(Address.From(Host.Of("localhost"), _testPort, AddressType.None),
+            _client = new BasicClientRequestResponseChannel(Address.From(Host.Of("localhost"), _testPort, AddressType.None),
                 _clientConsumer, PoolSize, 10240, logger);
 
             ++_testPort;
